changement type de retour de la methode permettant de générer le numeroCommande et quelques refactoring

+++ src/main/java/com/library/entities/CommandeClient.java
@@ -1,16 +1,14 @@
 package com.library.entities;
 
-import java.time.LocalDate;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonGetter;
 
+import javax.persistence.*;
+import javax.validation.Valid;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-
-import javax.persistence.*;
-import javax.validation.Valid;
-
-import com.fasterxml.jackson.annotation.*;
 
 @Entity
 @Table(name = "commandeClient")
@@ -28,7 +26,7 @@ public class CommandeClient implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private int numeroCommande;
+    private long numeroCommande;
 
     /*@DateTimeFormat(pattern = "yyyy-MM-dd")
     @Temporal(TemporalType.DATE)
@@ -55,7 +53,7 @@ public class CommandeClient implements Serializable {
         super();
     }
 
-    public CommandeClient(Long id, int numeroCommande, Date dateCommande, Client client, @Valid List<LigneCmdClient> lcomms, double totalCommande, String status) {
+    public CommandeClient(Long id, long numeroCommande, Date dateCommande, Client client, @Valid List<LigneCmdClient> lcomms, double totalCommande, String status) {
         this.id = id;
         this.numeroCommande = numeroCommande;
         this.dateCommande = dateCommande;
@@ -73,11 +71,11 @@ public class CommandeClient implements Serializable {
         this.id = id;
     }
 
-    public int getNumeroCommande() {
+    public long getNumeroCommande() {
         return numeroCommande;
     }
 
-    public void setNumeroCommande(int numeroCommande) {
+    public void setNumeroCommande(long numeroCommande) {
         this.numeroCommande = numeroCommande;
     }
