+++ src/main/java/com/library/repository/LigneCmdClientRepository.java
@@ -17,11 +17,11 @@ import javax.transaction.Transactional;
 @Repository
 public interface LigneCmdClientRepository extends JpaRepository<LigneCmdClient, Long> {
 
-	List<LigneCmdClient> findAllByNumero(int numero);
+	List<LigneCmdClient> findAllByNumero(long numero);
 
 	@Modifying
 	@Query("delete from LigneCmdClient where numero = :numero")
-	void deleteByNumero(@Param("numero") int numero);
+	void deleteByNumero(@Param("numero") long numero);
 
 	@Query("select p from LigneCmdClient p where p.produit.id =:prod")
 	public List<LigneCmdClient> ListLigneCmdClientByProduitId(@Param("prod") Long prodId);
